inbox_insights/requirements.txt
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
google-api-python-client>=2.0.0
pandas>=1.5.0
tqdm>=4.65.0
python-dotenv>=0.19.0
beautifulsoup4>=4.9.0
numpy>=1.20.0